virtual clients: switch to VLByteVec

+++ openmls/tests/virtual_clients.rs
@@ -1993,7 +1993,7 @@ fn vc_batch_key_packages_join_in_any_order() {
         .expect("alice_a build_vc_batch");
     let generation = batch.generation;
     assert_eq!(generation, 0, "the batch consumes a single generation");
-    assert_eq!(batch.key_packages.len(), count as usize);
+    assert_eq!(batch.key_packages.len(), count);
 
     let kp_infos = batch
         .key_packages
@@ -2017,10 +2017,7 @@ fn vc_batch_key_packages_join_in_any_order() {
 
     // The sibling joins via a HIGH batch index first and a LOW one second,
     // each through a separate higher-level group.
-    let high_bundle = batch.key_packages[(count - 1) as usize]
-        .0
-        .key_package()
-        .clone();
+    let high_bundle = batch.key_packages[count - 1].0.key_package().clone();
     let low_bundle = batch.key_packages[0].0.key_package().clone();
 
     for (label, kp) in [
